+++ pkg/sql/rowexec/processors_test.go
@@ -318,7 +318,7 @@ func TestProcessorBaseContext(t *testing.T) {
 	defer log.Scope(t).Close(t)
 
 	// Use a custom context to distinguish it from the background one.
-	ctx := context.WithValue(context.Background(), struct{}{}, struct{}{})
+	ctx := context.WithValue(context.Background(), contextKey{}, struct{}{})
 	st := cluster.MakeTestingClusterSettings()
 
 	runTest := func(t *testing.T, f func(noop *noopProcessor)) {
@@ -934,3 +934,5 @@ func testReaderProcessorDrain(
 		}
 	})
 }
+
+type contextKey struct{}
