+++ pkg/util/sysutil/socket.go
@@ -14,22 +14,6 @@ import (
 	"github.com/cockroachdb/errors"
 )
 
-// SetKeepAliveCount sets the keep alive probe count on a TCP
-// connection.
-func SetKeepAliveCount(conn *net.TCPConn, probeCount int) (err error) {
-	syscallConn, err := conn.SyscallConn()
-	if err != nil {
-		return err
-	}
-	outerErr := syscallConn.Control(func(fd uintptr) {
-		err = syscall.SetsockoptInt(SocketFd(fd), syscall.IPPROTO_TCP, syscall.TCP_KEEPCNT, probeCount)
-	})
-	if err != nil || outerErr != nil {
-		return errors.WithSecondaryError(err, outerErr)
-	}
-	return nil
-}
-
 // GetKeepAliveSettings gets the keep alive socket connections
 // set on a TCP connection.
 func GetKeepAliveSettings(
